feat: add install script for dotfiles
The install script creates symbolic links for all files in the .config and bin directories to the corresponding directories in the user's home directory. It also creates symbolic links for .bash_aliases, .bash_functions, .profile, .xinitrc, and .Xmodmap. The script can be run with the -f option to overwrite existing files.